requirements.txt
numpy
re
django_crontab
APScheduler==3.6.3
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
Deprecated==1.2.10
Django==3.1.3
gunicorn==20.0.4
idna==2.10
isort==5.6.4
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pymisp==2.4.134
pymongo==3.11.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.1
wrapt==1.12.1
